refactor(geo): extract order_eval to dsl/eval module, fix defensive copy and README
Move order_eval and validate_weighted_expressions! to DSL::Eval for proper separation
of concerns. Add .dup to GeoDistance#geo_distance_meters. Fix README model declaration.
